// src/enum/column-type.enum.ts
export enum ColumnType {
  BIGINT = 'bigint',
  BIGSERIAL = 'bigserial',
  BIT = 'bit',
  BIT_VARYING = 'bit varying',
  BOOLEAN = 'boolean',
  BOX = 'box',
  BYTEA = 'bytea',
  CHAR = 'character',
  CHAR_VARYING = 'character varying',
  CIDR = 'cidr',
  CIRCLE = 'circle',
  DATE = 'date',
  DOUBLE = 'double precision',
  INET = 'inet',
  INTEGER = 'integer',
  INTERVAL = 'interval',
  JSON = 'json',
  JSONB = 'jsonb',
  LINE = 'line',
  LSEG = 'lseg',
  MACADDR = 'macaddr',
  MACADDR8 = 'macaddr8',
  MONEY = 'money',
  NUMERIC = 'numeric',
  PATH = 'path',
  PG_LSN = 'pg_lsn',
  PG_SNAPSHOT = 'pg_snapshot',
  POINT = 'point',
  POLYGON = 'polygon',
  REAL = 'real',
  SMALLINT = 'smallint',
  SMALLSERIAL = 'smallserial',
  SERIAL = 'serial',
  TEXT = 'text',
  TIME = 'time without time zone',
  TIME_WITH_TIMEZONE = 'time with time zone',
  TIMESTAMP = 'timestamp without time zone',
  TIMESTAMP_WITH_TIMEZONE = 'timestamp with time zone',
  TSQUERY = 'tsquery',
  TSVECTOR = 'tsvector',
  TXID_SNAPSHOT = 'txid_snapshot',
  UUID = 'uuid',
  XML = 'xml',
}